pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown_vector"
version = "0.1.0"
description = "Markdown 文件向量化服務"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "gradio>=3.50.2",
    "chromadb>=0.4.18",
    "openai>=1.5.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.1",
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests.py", "integration_tests.py"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
